channel override bug! my soul hurts :c
