+++ .gitignore
@@ -1,2 +1,3 @@
 /target
 /s.json
+/data.json
