+++ Cargo.toml
@@ -8,7 +8,7 @@ edition = "2024"
 once_cell = "1.19"
 clap = { version = "4.5.47", features = ["derive"] }
 log = "0.4.28"
-poise = "0.6.1"
+poise = { version = "0.6.1", features = ["cache"] }
 serde = "1.0.225"
 simple_logger = "5.0.0"
 tokio = { version = "1.21.2", features = ["macros", "rt-multi-thread"] }
