+++ data.json
@@ -1,1 +1,1 @@
-{"db":{"440880669535633408":{"active_mimic":{"name":"Got To Make A Move To A Town That's Right For Me","avatar_url":"https://cdn.discordapp.com/ephemeral-attachments/1426728206056292465/1429979305496154140/export202510201946003390.png?ex=68f81b82&is=68f6ca02&hm=2fa26a5b2116c81e30d1da0c6915363def88c54043a91f4744c9b52447029cc7&"},"mimics":[{"name":"💸","avatar_url":null},{"name":"Got To Make A Move To A Town That's Right For Me","avatar_url":"https://cdn.discordapp.com/ephemeral-attachments/1426728206056292465/1429979305496154140/export202510201946003390.png?ex=68f81b82&is=68f6ca02&hm=2fa26a5b2116c81e30d1da0c6915363def88c54043a91f4744c9b52447029cc7&"}],"auto_mode":null},"326389428593623040":{"active_mimic":{"name":"asdf","avatar_url":null},"mimics":[{"name":"Maeson","avatar_url":"https://cdn.discordapp.com/avatars/330156315039105025/c0c54fa39d3c32ae2cc2abcd0ee493ed.webp?size=1024"},{"name":"asdf","avatar_url":null}],"auto_mode":true},"212986802481397761":{"active_mimic":{"name":"test","avatar_url":null},"mimics":[{"name":"test","avatar_url":null}],"auto_mode":null}}}
+{"db":{"440880669535633408":{"active_mimic":{"name":"💸","avatar_url":null},"mimics":[{"name":"💸","avatar_url":null},{"name":"Got To Make A Move To A Town That's Right For Me","avatar_url":"https://cdn.discordapp.com/ephemeral-attachments/1426728206056292465/1429979305496154140/export202510201946003390.png?ex=68f81b82&is=68f6ca02&hm=2fa26a5b2116c81e30d1da0c6915363def88c54043a91f4744c9b52447029cc7&"}],"auto_mode":false,"channel_override":{}},"1430263179618811935":{"active_mimic":null,"mimics":[],"auto_mode":null,"channel_override":{}},"1417653924139241584":{"active_mimic":null,"mimics":[],"auto_mode":null,"channel_override":{}},"1428890688044531723":{"active_mimic":null,"mimics":[],"auto_mode":null,"channel_override":{}},"326389428593623040":{"active_mimic":{"name":"asdf","avatar_url":null},"mimics":[{"name":"Maeson","avatar_url":"https://cdn.discordapp.com/avatars/330156315039105025/c0c54fa39d3c32ae2cc2abcd0ee493ed.webp?size=1024"},{"name":"asdf","avatar_url":null},{"name":"Static","avatar_url":"https://cdn.discordapp.com/ephemeral-attachments/1426728206056292465/1430262678152020068/image.png?ex=68f9236b&is=68f7d1eb&hm=e0ed8adb45f870c8c415f1716cd1abdab08601472aa5628cf6de8a1e03aa64c5&"}],"auto_mode":false,"channel_override":{}},"1430287773008396289":{"active_mimic":null,"mimics":[],"auto_mode":null,"channel_override":{}},"127227279104278528":{"active_mimic":null,"mimics":[],"auto_mode":null,"channel_override":{}},"212986802481397761":{"active_mimic":{"name":"test","avatar_url":null},"mimics":[{"name":"test","avatar_url":null}],"auto_mode":null,"channel_override":{}}}}
