+++ src/commands/mod.rs
@@ -5,6 +5,7 @@ use crate::{
     types::{Context, Data, EmbedType, Error, Reply},
     utils,
 };
+
 mod mimic;
 mod oot;
 mod vox;
