+++ src/mimic_db.rs
@@ -1,6 +1,4 @@
-use crate::types::{Mimic, MimicUser, Result};
-use poise::serenity_prelude::Webhook;
-use poise::serenity_prelude::{self as serenity};
+use crate::types::{Mimic, MimicUser};
 
 impl MimicUser {
     /// adds this Mimic to the mimics member variable of this user's MimicUser struct.
@@ -9,22 +7,4 @@ impl MimicUser {
     }
 }
 
-pub async fn get_or_create_webhook(
-    http: &serenity::Http,
-    channel_id: serenity::ChannelId,
-) -> Result<Webhook> {
-    const WEBHOOK_NAME: &str = "logosV3-mimic";
-    if let Ok(existing) = channel_id.webhooks(http).await
-        && let Some(w) = existing
-            .into_iter()
-            .find(|w| w.name.as_deref() == Some(WEBHOOK_NAME))
-    {
-        return Ok(w);
-    }
-
-    //the webby don't exist :c
-    let hook = channel_id
-        .create_webhook(http, serenity::CreateWebhook::new(WEBHOOK_NAME))
-        .await?;
-    Ok(hook)
-}
+mod modname {}
